% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{diffFinder}
\alias{diffFinder}
\title{Find the difference between landmark ages}
\usage{
diffFinder(age.vector)
}
\arguments{
\item{age.vector}{A vector of ages}
}
\value{
A list. Diffsame is a logical value indicating if the differences
between the landmark ages are all the same. diff provides the difference
between the elements
}
\description{
Checks if a vector of landmark ages is regularly spaced and
the distance between elements
}
